added ant media stream

+++ src/app/services/liveStreamingService.js
@@ -4,6 +4,7 @@ const { ObjectId } = require("mongodb");
 const { findOne } = require("../models/liveStreamModel");
 const liveStreamModel = require("../models/liveStreamModel");
 const socialMediaStreamingModel = require("../models/socialMediaStreamingModel");
+const { config } = require("dotenv");
 
 module.exports = {
   verify: async function (req) {
@@ -66,41 +67,33 @@ module.exports = {
     }
   },
   createLiveStream: async function (req) {
+    console.log('inside create live stream', req.body);
     try {
-      config = {
+      let config = {
         method: "POST",
-        url: "https://api.mux.com/video/v1/live-streams",
+        url: `${process.env.ANT_MEDIA_URL}broadcasts/create`,
         headers: {
           "Content-Type": "application/json",
-          Authorization:
-            "Basic " +
-            Buffer.from(
-              `${process.env.MUX_TOKEN_ID}:${process.env.MUX_TOKEN_SECRET}`
-            ).toString("base64"),
+          Authorization: `${process.env.ANT_MEDIA_AUTH}`
         },
-        data: {
-          playback_policy: ["public"],
-          new_asset_settings: {
-            playback_policy: ["public"],
-          },
-        },
+        data: req.body,
       };
-
       const muxData = await axios(config);
       console.log("mux mux mux", muxData.data);
       if (muxData) {
         const saveMuxData = await liveStreamModel.create({
-          streamKey: muxData.data.data.stream_key,
+          streamKey: muxData.data.streamId,
           createdBy: req.user.userId,
-          status: muxData.data.data.status,
-          muxStreamingId: muxData.data.data.id,
-          startTime: muxData.data.created_at,
-          playbackId: muxData.data.data.playback_ids[0].id
+          status: muxData.data.status,
+          muxStreamingId: "",
+          startTime: "",
+          playbackId: "",
         });
 
         return { status: true, data: saveMuxData };
       }
     } catch (error) {
+      console.log(error.response.data)
       return { status: false, code: 500, msg: `${error.message}` };
     }
   },
@@ -109,24 +102,36 @@ module.exports = {
 
     const youtube = req.body.youtube;
     try {
-      let config = {
-        method: "POST",
-        url: `https://api.mux.com/video/v1/live-streams/${liveStreamId}/simulcast-targets`,
-        headers: {
-          "Content-Type": "application/json",
-          Authorization:
-            "Basic " +
-            Buffer.from(
-              `${process.env.MUX_TOKEN_ID}:${process.env.MUX_TOKEN_SECRET}`
-            ).toString("base64"),
-        },
-        data: {
-          url: youtube.url,
-          stream_key: youtube.streamKey,
-          passthrough: youtube.passthrough,
-        },
-      };
-      const streamData = await axios(config);
+      let configForAnt = {
+
+          method: "POST",
+          url: `${process.env.ANT_MEDIA_URL}broadcasts/${liveStreamId}/rtmp-endpoint`,
+          headers: {
+            "Content-Type": "application/json",
+            Authorization:`${process.env.ANT_MEDIA_AUTH}`
+          },
+          data: {
+            "rtmpUrl":youtube.url
+          },
+      }
+      // let config = {
+      //   method: "POST",
+      //   url: `${process.env.ANT_MEDIA_URL}broadcasts/${liveStreamId}`,
+      //   headers: {
+      //     "Content-Type": "application/json",
+      //     Authorization:
+      //       "Basic " +
+      //       Buffer.from(
+      //         `${process.env.MUX_TOKEN_ID}:${process.env.MUX_TOKEN_SECRET}`
+      //       ).toString("base64"),
+      //   },
+      //   data: {
+      //     url: youtube.url,
+      //     stream_key: youtube.streamKey,
+      //     passthrough: youtube.passthrough,
+      //   },
+      // };
+      const streamData = await axios(configForAnt);
       if (streamData) {
         let livestreamData = await liveStreamModel.findOne({
           muxStreamingId: liveStreamId,
@@ -134,11 +139,11 @@ module.exports = {
         const simulcastData = await socialMediaStreamingModel.create({
           liveStreamId: ObjectId(livestreamData._id),
           platformName: "youtube",
-          muxSimulcastId: streamData.data.data.id,
-          url: streamData.data.data.url,
-          streamKey: streamData.data.data.streamKey,
-          status: streamData.data.data.status,
-          passthrough: streamData.data.data.passthrough,
+          muxSimulcastId: "",
+          url: youtube.url,
+          streamKey: youtube.streamKey,
+          status: "",
+          passthrough: youtube && youtube.passthrough ?  youtube.passthrough : '',
           createdBy: req.user.userId,
         });
         livestreamData.socialMediaIds.push(simulcastData._id);
@@ -158,36 +163,61 @@ module.exports = {
     const liveStreamId = req.body.liveStreamId;
     const twitch = req.body.twitch;
     try {
-      config = {
+      let configForAnt = {
+
         method: "POST",
-        url: `https://api.mux.com/video/v1/live-streams/${liveStreamId}/simulcast-targets`,
+        url: `${process.env.ANT_MEDIA_URL}broadcasts/${liveStreamId}/rtmp-endpoint`,
         headers: {
           "Content-Type": "application/json",
-          Authorization:
-            "Basic " +
-            Buffer.from(
-              `${process.env.MUX_TOKEN_ID}:${process.env.MUX_TOKEN_SECRET}`
-            ).toString("base64"),
+          Authorization:`${process.env.ANT_MEDIA_AUTH}`
         },
         data: {
-          url: "rtmp://live.twitch.tv/app/",
-          stream_key: twitch.streamKey,
-          passthrough: twitch.passthrough,
+          "rtmpUrl":twitch.url
         },
-      };
-      const streamData = await axios(config);
+    }
+      // config = {
+      //   method: "POST",
+      //   url: `https://api.mux.com/video/v1/live-streams/${liveStreamId}/simulcast-targets`,
+      //   headers: {
+      //     "Content-Type": "application/json",
+      //     Authorization:
+      //       "Basic " +
+      //       Buffer.from(
+      //         `${process.env.MUX_TOKEN_ID}:${process.env.MUX_TOKEN_SECRET}`
+      //       ).toString("base64"),
+      //   },
+      //   data: {
+      //     url: twitch.url,
+      //     stream_key: twitch.streamKey,
+      //     passthrough: twitch.passthrough,
+      //   },
+      // };
+      const streamData = await axios(configForAnt);
       if (streamData) {
         let livestreamData = await liveStreamModel.findOne({
           muxStreamingId: liveStreamId,
         });
+        // const simulcastData = await socialMediaStreamingModel.create({
+        //   liveStreamId: ObjectId(livestreamData._id),
+        //   platformName: "twitch",
+        //   muxSimulcastId: streamData.data.data.id,
+        //   url: streamData.data.data.url,
+        //   streamKey: streamData.data.data.streamKey,
+        //   status: streamData.data.data.status,
+        //   passthrough: streamData.data.data.passthrough,
+        //   createdBy: req.user.userId,
+        // });
+        // livestreamData.socialMediaIds.push(simulcastData._id);
+        // const saved = await livestreamData.save();
+        // return { status: true, data: streamData.data.data };
         const simulcastData = await socialMediaStreamingModel.create({
           liveStreamId: ObjectId(livestreamData._id),
           platformName: "twitch",
-          muxSimulcastId: streamData.data.data.id,
-          url: streamData.data.data.url,
-          streamKey: streamData.data.data.streamKey,
-          status: streamData.data.data.status,
-          passthrough: streamData.data.data.passthrough,
+          muxSimulcastId: "",
+          url: twitch.url,
+          streamKey: twitch.streamKey,
+          status: "",
+          passthrough: twitch && twitch.passthrough ?  twitch.passthrough : '',
           createdBy: req.user.userId,
         });
         livestreamData.socialMediaIds.push(simulcastData._id);
@@ -207,36 +237,58 @@ module.exports = {
     const liveStreamId = req.body.liveStreamId;
     const facebook = req.body.facebook;
     try {
-      config = {
+      // config = {
+      //   method: "POST",
+      //   url: `https://api.mux.com/video/v1/live-streams/${liveStreamId}/simulcast-targets`,
+      //   headers: {
+      //     "Content-Type": "application/json",
+      //     Authorization:
+      //       "Basic " +
+      //       Buffer.from(
+      //         `${process.env.MUX_TOKEN_ID}:${process.env.MUX_TOKEN_SECRET}`
+      //       ).toString("base64"),
+      //   },
+      //   data: {
+      //     url: facebook.url,
+      //     stream_key: facebook.streamKey,
+      //     passthrough: facebook.passthrough,
+      //   },
+      // };
+      let configForAnt = {
+
         method: "POST",
-        url: `https://api.mux.com/video/v1/live-streams/${liveStreamId}/simulcast-targets`,
+        url: `${process.env.ANT_MEDIA_URL}broadcasts/${liveStreamId}/rtmp-endpoint`,
         headers: {
           "Content-Type": "application/json",
-          Authorization:
-            "Basic " +
-            Buffer.from(
-              `${process.env.MUX_TOKEN_ID}:${process.env.MUX_TOKEN_SECRET}`
-            ).toString("base64"),
+          Authorization:`${process.env.ANT_MEDIA_AUTH}`
         },
         data: {
-          url: "rtmps://live-api-s.facebook.com:443/rtmp/",
-          stream_key: facebook.streamKey,
-          passthrough: facebook.passthrough,
+          "rtmpUrl":facebook.url
         },
-      };
-      const streamData = await axios(config);
+    }
+      const streamData = await axios(configForAnt);
       if (streamData) {
         let livestreamData = await liveStreamModel.findOne({
           muxStreamingId: liveStreamId,
         });
+        // const simulcastData = await socialMediaStreamingModel.create({
+        //   liveStreamId: ObjectId(livestreamData._id),
+        //   platformName: "facebook",
+        //   muxSimulcastId: streamData.data.data.id,
+        //   url: streamData.data.data.url,
+        //   streamKey: streamData.data.data.streamKey,
+        //   status: streamData.data.data.status,
+        //   passthrough: streamData.data.data.passthrough,
+        //   createdBy: req.user.userId,
+        // });
         const simulcastData = await socialMediaStreamingModel.create({
           liveStreamId: ObjectId(livestreamData._id),
           platformName: "facebook",
-          muxSimulcastId: streamData.data.data.id,
-          url: streamData.data.data.url,
-          streamKey: streamData.data.data.streamKey,
-          status: streamData.data.data.status,
-          passthrough: streamData.data.data.passthrough,
+          muxSimulcastId: "",
+          url: facebook.url,
+          streamKey: facebook.streamKey,
+          status: "",
+          passthrough: facebook && facebook.passthrough ?  facebook.passthrough : '',
           createdBy: req.user.userId,
         });
 
@@ -259,7 +311,7 @@ module.exports = {
       const liveStreamId = req.body.muxStreamingId;
       const socialMediaStream = req.body.muxSimulcastId;
       console.log('live stream', liveStreamId, 'live stream 00', socialMediaStream);
-      config = {
+      let config = {
         method: "DELETE",
         url: `https://api.mux.com/video/v1/live-streams/${liveStreamId}/simulcast-targets/${socialMediaStream}`,
         headers: {
@@ -328,22 +380,34 @@ module.exports = {
     try {
       const liveStreamId = req.query.id;
       const findLiveStream = await liveStreamModel.findOne({_id:ObjectId(liveStreamId)});
+        let configForAnt = {
 
-      config = {
-        method: "PUT",
-        url: `https://api.mux.com/video/v1/live-streams/${liveStreamId}/complete`,
-        headers: {
-          "Content-Type": "application/json",
-          Authorization:
-            "Basic " +
-            Buffer.from(
-              `${process.env.MUX_TOKEN_ID}:${process.env.MUX_TOKEN_SECRET}`
-            ).toString("base64"),
-        },
-      };
-      const muxData = await axios(config);
+          method: "PUT",
+          url: `${process.env.ANT_MEDIA_URL}broadcasts/${liveStreamId}`,
+          headers: {
+            "Content-Type": "application/json",
+            Authorization:`${process.env.ANT_MEDIA_AUTH}`
+          },
+          data: {
+            "status": "finished",
+            "playListStatus": "finished",
+            },
+        }
+      // let config = {
+      //   method: "PUT",
+      //   url: `https://api.mux.com/video/v1/live-streams/${liveStreamId}/complete`,
+      //   headers: {
+      //     "Content-Type": "application/json",
+      //     Authorization:
+      //       "Basic " +
+      //       Buffer.from(
+      //         `${process.env.MUX_TOKEN_ID}:${process.env.MUX_TOKEN_SECRET}`
+      //       ).toString("base64"),
+      //   },
+      // };
+      const muxData = await axios(configForAnt);
       if(muxData){
-        findLiveStream.status = 'Completed';  
+        findLiveStream.status = 'finished';  
         await findLiveStream.save();   
         return {status:true, msg:"Live stream marked Completed !"}
       }
@@ -354,33 +418,31 @@ module.exports = {
   deleteLiveStream: async function (req){
     try {
       const userId = req.user.userId;
-      const data =  await liveStreamModel.find({createdBy:ObjectId(userId)});
-
-      for (let i = 0; i < data.length; i++) {
-        const element = data[i].muxStreamingId;
-        console.log('element element', element);
-        console.log('elelement.muxStreamingId', element);
-        config = {
-          method: "DELETE",
-          url: `https://api.mux.com/video/v1/live-streams/${element}`,
-          headers: {
-            "Content-Type": "application/json",
-            Authorization:
-              "Basic " +
-              Buffer.from(
-                `${process.env.MUX_TOKEN_ID}:${process.env.MUX_TOKEN_SECRET}`
-              ).toString("base64"),
-          },
-        };
-        console.log('config', config);
-        const streamData = await axios(config);
+     const deleteLiveStreamID = req.params.liveStreamId;
+     if(!deleteLiveStreamID){
+      return {status:false, code:400,msg:"No stream Id found in params"}
+     }
+      
+      let configForAnt = {
+        method: "DELETE",
+        url: `${process.env.ANT_MEDIA_URL}broadcasts/${liveStreamId}/${deleteLiveStreamID}`,
+        headers: {
+          "Content-Type": "application/json",
+          Authorization:`${process.env.ANT_MEDIA_AUTH}`
+        },
+      }
+      const findStream = await liveStreamModel.findOne({liveStreamId:deleteLiveStreamID});
+        findStream.isDeleted = true;
+      const savedSaved = await findStream.save()
+        console.log('config', savedSaved);
+        const streamData = await axios(configForAnt);
         console.log('stream Data stream Data',streamData)
+      if(streamData.data){
+        return {status:false, code:200, msg:"Stream deleted successfully!"}
       }
     } catch (error) {
       console.log('ee',error )
       return {status: false, code :500, msg:`${error.message}`}
     }
-
-
   }
 };
